# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AbIratoBooks API",
        "description": "AbIratoBooks API",
        "contact": {
            "email": "amandine.pda@gmx.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/statuses": {
            "get": {
                "tags": [
                    "Status"
                ],
                "summary": "Get all status",
                "operationId": "e96ea40c6f497546795df9ec612e1928",
                "responses": {
                    "200": {
                        "description": "List of status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Status"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/statuses/{id}": {
            "get": {
                "tags": [
                    "Status"
                ],
                "summary": "Get a status by ID",
                "operationId": "ea554faff3dfcadd142b14010e8b3331",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Status"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Status not found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/admin-only/statuses": {
            "post": {
                "tags": [
                    "Status"
                ],
                "summary": "Create a new status",
                "operationId": "548cbc768800dec461afd28fc5c48381",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Status"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The created status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Status"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/admin-only/statuses/{id}": {
            "put": {
                "tags": [
                    "Status"
                ],
                "summary": "Update a status",
                "operationId": "3a3e4e2c2f5272415da32e0808e7db3b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Status"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The updated status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Status"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Status not found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Status"
                ],
                "summary": "Delete a status",
                "operationId": "45d600bd4f5c544b3c9b62660f1987ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Status not found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Status": {
                "title": "Status",
                "description": "Status model",
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "description": "Status value",
                        "type": "string",
                        "example": "Vendu"
                    },
                    "created_at": {
                        "description": "Creation date and time",
                        "type": "string",
                        "format": "datetime",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Last update date and time",
                        "type": "string",
                        "format": "datetime",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}